requirements.txt
# Dependências essenciais
python-dotenv==1.1.0
groq==0.25.0

# Dependências opcionais para futuras expansões
requests==2.32.3
